Create Multiomic_RF_pipeline6
This pipeline ingests several different omic data sets. Merges the data with metadata so that it is ready for model building. 
A random forest model is used to prioritize features and make predictions. K-fold validation is used to robustly calculate performance and feature importance. The pipeline removes colinear features with low importance scores. Once this is completed the next rounds of feature selection are performed by removing the bottom 10% of features until only one is left. 
At each iteration, the average performance, SD, and signal/noise are calculated across folds for both performance and feature importance. The results are then summarized at the end. Once one data set has been reduced to one feature a new omic set is loaded into the pipeline.